feat: Add Dockerfile and docker-compose.yml
Add Dockerfile and docker-compose.yml to enable containerization and deployment of the frontend application. The Dockerfile defines the build process and sets up the necessary dependencies, while the docker-compose.yml file configures the frontend service and sets up the network. This allows the application to be easily deployed and run in a containerized environment.